requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
awsebcli==3.20.10
blessed==1.20.0
blinker==1.8.2
botocore==1.31.85
cachetools==5.4.0
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
macholib==1.16.3
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
openai==1.39.0
packaging==24.1
pandas==2.0.3
pathspec==0.10.1
playwright==1.45.1
playwright-stealth==1.0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.1
termcolor==1.1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.1.9
Werkzeug==3.0.3
zipp==3.19.2
